requirements.txt
# Core
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# OpenRouter API
openai==1.6.1
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Code metrics
radon==6.0.1
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-java>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-cpp>=0.21.0

# Tree edit distance
apted==1.0.3
zss==1.2.0

# Web scraping
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Analysis
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Utilities
loguru==0.7.2
tqdm==4.66.1
rich==13.7.0
